guy's comments
