+++ packages/contract_harness/Cargo.toml
@@ -28,10 +28,6 @@ treasury_manager = ["dep:treasury_manager"]
 shade-oracles = ["dep:shade-oracles"]
 shade-oracles-ensemble = ["dep:shade-oracles-ensemble"]
 sky = ["dep:sky"]
-#shadeswap_exchange = ["dep:amm_pair"]
-#shadeswap_factory = ["dep:factory"]
-#mock_shdswp = ["dep:mock-shade-pair"]
-#shadeswap_lp_token = ["dep:lp-token"]
 
 [dependencies]
 cosmwasm-std = { version = "0.10", package = "secret-cosmwasm-std" }
@@ -39,11 +35,6 @@ fadroma = { branch = "v100", commit = 76867e0, git = "https://github.com/hackbg/
   "ensemble",
 ] }
 sky = { version = "0.1.0", path = "../../contracts/sky", optional = true }
-#fadroma = { branch = "staking", git = "https://github.com/a-stgeorge/fadroma-staking" }
-#amm_pair = { git = "https://github.com/securesecrets/shadeswap", optional = true }
-#factory = { git = "https://github.com/securesecrets/shadeswap", optional = true }
-#mock-shade-pair = { branch = "shadeswap-market-oracle", git = "https://github.com/securesecrets/shade-oracle", optional = true }
-#lp-token = { git = "https://github.com/securesecrets/shadeswap", optional = true }
 mint = { version = "0.1.0", path = "../../contracts/mint", optional = true }
 oracle = { version = "0.1.0", path = "../../contracts/oracle", optional = true }
 mock_band = { version = "0.1.0", path = "../../contracts/mock_band", optional = true }
