+++ packages/contract_harness/src/harness_macro.rs
@@ -24,31 +24,3 @@ macro_rules! implement_harness {
 }
 
 pub(crate) use implement_harness;
-
-macro_rules! implement_harness_fadroma {
-    ($x:ident, $s:ident) => {
-        use cosmwasm_std::{InitResponse};
-        use fadroma::ensemble::{ContractHarness, MockDeps};
-        use fadroma::scrt::cosmwasm_std::{from_binary, Binary, Env, HandleResponse, StdResult};
-        impl ContractHarness for $x {
-            fn init(&self, deps: &mut MockDeps, env: Env, msg: Binary) -> StdResult<InitResponse> {
-                $s::contract::init(deps, env, from_binary(&msg)?)
-            }
-
-            fn handle(
-                &self,
-                deps: &mut MockDeps,
-                env: Env,
-                msg: Binary,
-            ) -> StdResult<HandleResponse> {
-                $s::contract::handle(deps, env, from_binary(&msg)?)
-            }
-
-            fn query(&self, deps: &MockDeps, msg: Binary) -> StdResult<Binary> {
-                $s::contract::query(deps, from_binary(&msg)?)
-            }
-        }
-    };
-}
-
-pub(crate) use implement_harness_fadroma;
