+++ packages/shade_protocol/Cargo.toml
@@ -18,7 +18,7 @@ default = ["utils"]
 # TODO: Normalize usage, some features are using - while others use _
 
 # Templates
-dex = ["utils", "math", "snip20", "mint", "band", "dep:fadroma"]
+dex = ["utils", "math", "snip20", "mint", "band"]
 band = []
 secretswap = ["utils"]
 sienna = ["utils", "math"]
@@ -50,14 +50,14 @@ adapter = []
 snip20 = ["utils", "errors", "dep:base64", "query_auth_lib"]
 query_auth = ["utils", "query_auth_lib", "dep:remain"]
 snip20_staking = ["utils", "storage"]
-sky = ["mint", "snip20", "utils", "dex", "sienna", "secretswap", "shadeswap", "adapter", "storage_plus", "dep:secret-storage-plus"]
+sky = ["utils", "adapter", "storage_plus"]
 
 # Protocol Implementation Contracts
 # Used in internal smart contracts
 governance-impl = ["governance", "storage"]
 snip20-impl = ["snip20", "storage_plus"]
 query_auth_impl = ["query_auth", "storage_plus", "dep:base64"]
-#sky-impl = ["sky", "storage_plus"]
+sky-impl = ["sky", "dex", "secretswap", "sienna", "shadeswap", "mint"]
 
 # for quicker tests, cargo test --lib
 # for more explicit tests, cargo test --features=backtraces
